throw ogl portals into their own files
